requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
google-cloud-aiplatform>=1.35.0
vertexai>=0.0.1
unidecode>=1.3.6
google-auth>=2.23.0 